diverse Anpassungen wg 1000K

+++ prog/aufraeumen.sh
@@ -8,6 +8,8 @@ lokal=$(pwd | awk '{print $1}')
 source  ${lokal}/parfiles/steuerungsvariablen.ctr.sh
 ###############################################################
 
+
+
 ${BIN_DIR}/archivierung.sh
 
 echo "loesche files im $TMP_DIR"
@@ -19,27 +21,26 @@ mkdir -p $GCA_DIR
 echo "loesche files im $WORK_DIR"
 rm -rf $WORK_DIR
 mkdir -p $WORK_DIR
-#rm -f $SHP_DIR/*
 echo "loesche files im $FIM_DIR"
 rm -rf $FIM_DIR/*
 echo "loesche files im $SMS_DIR"
 rm -rf $SMS_DIR/*
 echo "loesche files im $LOG_DIR"
 rm -f $LOG_DIR/*
+if [ -d "${LAB_DIR}" ]; then
+   echo Loesche files im $LAB_DIR
+   ls -trl $LAB_DIR/
+   rm -f $LAB_DIR/*
+   ls -trl $LAB_DIR/
+fi
 
-echo Loesche files im $LAB_DIR
-ls -trl $LAB_DIR/
-rm -f $LAB_DIR/*
-ls -trl $LAB_DIR/
-
-
-#for dd in HDimputing FHDimputing; do 
-#echo "pruge now ${dd} direcctory"
-#cd ${SNP_DIR}/${dd}
-#prog/aufraeumen.sh
-#cd ${lokal}
-#echo " ";
-#done
+for dd in ${secondIMPUTATIONDIR}; do 
+  echo "pruge now ${dd} direcctory"
+  cd ${dd}
+  ${secondIMPUTATIONDIR}/prog/aufraeumen.sh
+  cd ${lokal}
+  echo " ";
+done
